pyproject.toml
[project]
name = "dc-video-mcp"
version = "0.1.0"
description = "Lightweight MCP server for video analysis: keyframe captioning, ASR, and semantic search."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "lgao", email = "lgao@fastspring.com" }]
dependencies = [
    "mcp>=1.0.0",
    "numpy>=1.26",
    "Pillow>=10.0",
    "sqlite-vec>=0.1.6",
    # Heavy ML deps: lazy-imported in handlers, but declared so `uv sync` installs them.
    "torch>=2.2",
    "transformers>=4.44,<5",
    "accelerate>=0.33",
    "einops>=0.8",
    "faster-whisper>=1.0",
    "sentence-transformers>=3.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[project.scripts]
dc-video-mcp = "dc_video_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dc_video_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
